requirements.txt
#Required modules
matplotlib
pyserial
